+++ src/main/java/subway/line/application/LineService.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,10 +8,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import subway.exception.ErrorMessage;
 import subway.exception.NotFoundException;
+import subway.line.domain.Line;
+import subway.line.domain.LineRepository;
 import subway.line.dto.LineRequest;
 import subway.line.dto.LineResponse;
-import subway.station.Station;
-import subway.station.StationRepository;
+import subway.station.domain.Station;
+import subway.station.domain.StationRepository;
 import subway.station.dto.StationResponse;
 
 @Service
