+++ src/main/java/subway/line/domain/LineRepository.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
