+++ src/main/java/subway/line/section/domain/SectionRepository.java
@@ -1,10 +1,10 @@
-package subway.line.section;
+package subway.line.section.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import subway.line.Line;
-import subway.station.Station;
+import subway.line.domain.Line;
+import subway.station.domain.Station;
 
 @Repository
 public interface SectionRepository extends JpaRepository<Section, Long> {
