+++ src/main/java/subway/station/domain/StationRepository.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
