+++ src/main/java/subway/station/web/StationController.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.web;
 
 import java.net.URI;
 import java.util.List;
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import subway.station.application.StationService;
 import subway.station.dto.StationRequest;
 import subway.station.dto.StationResponse;
